remove unused imports
No need to import anything we don't use in that file.
One reason is performance - why import stuff of scipy if we don't need it?
Another is cleanliness - we want to see in the code what depends on what. If everything just imports everything, it's hard to see where things come from.