Test files have been moved, include them in package
This is important in particular to make the CI run on GH.
Also, rewrite tests for make_lc so that the data files are found
both in an install and in a local development setup.
While we're at it, also change the test to compare more of the
data columns with reasonable uncertainties.